finalProj/requirements.txt
click==8.1.3
Flask==2.2.2
grpcio==1.51.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.11.0
